i74: fixed output of ack --count foo single-file
This fixes issue 74. When given a single file on the command line, ack no
longer outputs '<filename>:<count>' but simply '<count>'.

Furthermore, when --count is used together with -h (--no-filename), then a
single number is output, the total count over all files. grep outputs here a
list of numbers instead, which seems not very helpful.

WARNING: This checkin had to change the interface of App::Ack::print_count()
and App::Ack::print_count0(), both take an additional parameter (whether to
show the filename or not). So things like App::Wack need to adapt to this
change. The module test t/module.t was adjusted to cover these changes.